requirements.txt
# DigIdentity Engine — Python Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0

# Async
httpx>=0.27.0

# Validation
pydantic[email]>=2.0.0

# Database
supabase>=2.0.0

# Task Queue
celery[redis]>=5.4.0
redis>=5.0.0

# Payments
stripe>=8.0.0

# AI
anthropic>=0.40.0

# Scraping
requests>=2.32.0
beautifulsoup4>=4.12.0

# PDF Generation
weasyprint>=62.0
markdown>=3.7

# Email (stdlib — no extra deps needed)

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
